move things around to avoid circular imports
