+++ allennlp/training/trainer.py
@@ -12,8 +12,8 @@ import torch
 import torch.optim.lr_scheduler
 
 from allennlp.common import Params
-from allennlp.common.checks import ConfigurationError
-from allennlp.common.util import (dump_metrics, gpu_memory_mb, parse_cuda_device, peak_memory_mb,
+from allennlp.common.checks import ConfigurationError, parse_cuda_device
+from allennlp.common.util import (dump_metrics, gpu_memory_mb, peak_memory_mb,
                                   get_frozen_and_tunable_parameter_names, lazy_groups_of)
 from allennlp.common.tqdm import Tqdm
 from allennlp.data.instance import Instance
